[Refactor] presentation feature 별 모듈을 패키지로 변경
