+++ app/src/main/java/com/hgdgds/kokkok/navigation/KokNavHost.kt
@@ -3,11 +3,11 @@ package com.hgdgds.kokkok.navigation
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
-import com.hgdgds.home.navigation.HomeBase
-import com.hgdgds.home.navigation.homeSection
+import com.hgdgds.kokkok.mypage.navigation.myPageSection
+import com.hgdgds.kokkok.presentation.home.navigation.HomeBase
+import com.hgdgds.kokkok.presentation.home.navigation.homeSection
+import com.hgdgds.kokkok.presentation.reserve.navigation.reserveSection
 import com.hgdgds.kokkok.ui.KokAppState
-import com.hgdgds.mypage.navigation.myPageSection
-import com.hgdgds.reserve.navigation.reserveSection
 
 @Composable
 fun KokNavHost(
