--- presentation/home/src/main/java/com/hgdgds/home/HomeScreen.kt
@@ -1,12 +1,0 @@
-package com.hgdgds.home
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun HomeScreen() {
-    Column {
-        Text("Home")
-    }
-}
