--- presentation/home/src/test/java/com/hgdgds/home/ExampleUnitTest.kt
@@ -1,3 +1,0 @@
-package com.hgdgds.home
-
-class ExampleUnitTest
