--- presentation/mypage/src/main/java/com/hgdgds/mypage/MyPageScreen.kt
@@ -1,12 +1,0 @@
-package com.hgdgds.mypage
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun MyPageScreen() {
-    Column {
-        Text("MyPage")
-    }
-}
