--- presentation/reserve/src/main/java/com/hgdgds/reserve/ReserveScreen.kt
@@ -1,12 +1,0 @@
-package com.hgdgds.reserve
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun ReserveScreen() {
-    Column {
-        Text("Reserve")
-    }
-}
