--- presentation/reserve/src/test/java/com/hgdgds/reserve/ExampleUnitTest.kt
@@ -1,3 +1,0 @@
-package com.hgdgds.reserve
-
-class ExampleUnitTest
