+++ presentation/src/main/java/com/hgdgds/kokkok/presentation/home/HomeScreen.kt
@@ -1,0 +1,12 @@
+package com.hgdgds.kokkok.presentation.home
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun HomeScreen() {
+    Column {
+        Text("Home")
+    }
+}
