+++ presentation/src/main/java/com/hgdgds/kokkok/presentation/mypage/MyPageScreen.kt
@@ -1,0 +1,12 @@
+package com.hgdgds.kokkok.mypage
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MyPageScreen() {
+    Column {
+        Text("MyPage")
+    }
+}
