+++ presentation/src/main/java/com/hgdgds/kokkok/presentation/reserve/ReserveScreen.kt
@@ -1,0 +1,12 @@
+package com.hgdgds.kokkok.presentation.reserve
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun ReserveScreen() {
+    Column {
+        Text("Reserve")
+    }
+}
