+++ settings.gradle.kts
@@ -25,6 +25,4 @@ include(":core:ui")
 include(":core:designsystem")
 include(":data")
 include(":domain")
-include(":presentation:home")
-include(":presentation:reserve")
-include(":presentation:mypage")
+include(":presentation")
